+++ cmd/sfeth/cli/utils.go
@@ -23,6 +23,7 @@ import (
 	"github.com/lithammer/dedent"
 	"github.com/logrusorgru/aurora"
 	"github.com/spf13/viper"
+	"github.com/streamingfast/dstore"
 	"github.com/streamingfast/logging"
 	"go.uber.org/zap"
 )
@@ -40,9 +41,60 @@ func MustReplaceDataDir(dataDir, in string) string {
 	return in
 }
 
+var commonStoresCreated bool
+var indexStoreCreated bool
+
+func GetCommonStoresURLs(dataDir string) (mergedBlocksStoreURL, oneBlocksStoreURL, forkedBlocksStoreURL string, err error) {
+	mergedBlocksStoreURL = MustReplaceDataDir(dataDir, viper.GetString("common-blocks-store-url"))
+	oneBlocksStoreURL = MustReplaceDataDir(dataDir, viper.GetString("common-oneblock-store-url"))
+	forkedBlocksStoreURL = MustReplaceDataDir(dataDir, viper.GetString("common-forkedblocks-store-url"))
+
+	if commonStoresCreated {
+		return
+	}
+
+	if err = mkdirStorePathIfLocal(forkedBlocksStoreURL); err != nil {
+		return
+	}
+	if err = mkdirStorePathIfLocal(oneBlocksStoreURL); err != nil {
+		return
+	}
+	if err = mkdirStorePathIfLocal(mergedBlocksStoreURL); err != nil {
+		return
+	}
+	commonStoresCreated = true
+	return
+}
+
+func GetIndexStore(dataDir string) (indexStore dstore.Store, possibleIndexSizes []uint64, err error) {
+	indexStoreURL := viper.GetString("common-index-store-url")
+
+	if indexStoreURL != "" {
+		s, err := dstore.NewStore(indexStoreURL, "", "", false)
+		if err != nil {
+			return nil, nil, fmt.Errorf("couldn't create indexStore: %w", err)
+		}
+		if !indexStoreCreated {
+			if err = mkdirStorePathIfLocal(indexStoreURL); err != nil {
+				return nil, nil, err
+			}
+		}
+		indexStoreCreated = true
+		indexStore = s
+	}
+
+	for _, size := range viper.GetIntSlice("common-block-index-sizes") {
+		if size < 0 {
+			return nil, nil, fmt.Errorf("invalid negative size for common-block-index-sizes: %d", size)
+		}
+		possibleIndexSizes = append(possibleIndexSizes, uint64(size))
+	}
+	return
+}
+
 func mkdirStorePathIfLocal(storeURL string) (err error) {
-	zlog.Debug("creating directory and its parent(s)", zap.String("directory", storeURL))
 	if dirs := getDirsToMake(storeURL); len(dirs) > 0 {
+		zlog.Debug("creating directory and its parent(s)", zap.String("directory", storeURL))
 		err = makeDirs(dirs)
 	}
 	return